requirements.txt
django>=3,<4
celery[redis]
celery-redbeat
django-model-utils
django-webpack-loader>=1.0.0
django-js-reverse
django-import-export
djangorestframework
django-debreach
python-decouple
psycopg2
brotlipy
django-log-request-id
dj-database-url
gunicorn
whitenoise
psutil
ipython
sentry-sdk
pillow
djangorestframework-simplejwt
